tutunovka_bot/requirements.txt
telebot==0.0.5
python-dotenv==1.0.1
sqlalchemy==2.0.29